requirements.txt
mcstatus
pymongo
tzdata
python-dotenv